requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --resolver=backtracking
#
asgiref==3.6.0
    # via django
attrs==22.2.0
    # via
    #   cattrs
    #   ddtrace
    #   jsonschema
boto3==1.26.97
    # via -r requirements.in
botocore==1.29.97
    # via
    #   boto3
    #   s3transfer
bytecode==0.14.0
    # via ddtrace
cattrs==22.2.0
    # via ddtrace
ddsketch==2.0.4
    # via ddtrace
ddtrace==1.10.0
    # via -r requirements.in
django==4.1.7
    # via
    #   -r requirements.in
    #   django-storages
django-environ==0.10.0
    # via -r requirements.in
django-storages==1.13.2
    # via -r requirements.in
envier==0.4.0
    # via ddtrace
gevent==22.10.2
    # via gunicorn
greenlet==2.0.2
    # via gevent
gunicorn[gevent]==20.1.0
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.17.3
    # via ddtrace
packaging==23.0
    # via ddtrace
protobuf==4.22.1
    # via
    #   ddsketch
    #   ddtrace
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via botocore
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   ddsketch
    #   ddtrace
    #   python-dateutil
sqlparse==0.4.3
    # via django
tenacity==8.2.2
    # via ddtrace
typing-extensions==4.5.0
    # via ddtrace
urllib3==1.26.15
    # via botocore
whitenoise==6.4.0
    # via -r requirements.in
xmltodict==0.13.0
    # via ddtrace
zope-event==4.6
    # via gevent
zope-interface==6.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
